+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Removed
 - Removed generation of base/core files.
+- Removed `-module` option from `pywinrt.exe`.
 
 ## [v1.0.0-beta.10] - 2023-08-11